This commit was manufactured by cvs2svn to create branch 'jobmanager_history'.
Sprout from windows_2 2002-05-17 22:45:21 UTC cvs2git 'This commit was manufactured by cvs2svn to create branch 'windows_2'.'
Cherrypick from master 2002-06-03 16:41:26 UTC meder '.':
    common/source/dirt.sh
    common/source/library/globus_error.c
    common/source/library/globus_error.h
    core/setup/Makefile.am
    core/setup/dirt.sh
    core/setup/findshelltools.in
    core/setup/globus-makefile-header.in
    core/setup/globus-sh-tools-vars.sh.in
    core/setup/setup-globus-core.pl
    core/source/config/acthreads.m4
    core/source/dirt.sh
    core/source/scripts/Makefile.am
    core/source/scripts/globus-sh-tools.sh
    gsi/ssl_utils/source/dirt.sh
    gsi/ssl_utils/source/library/sslutils.c
    user_env/source/dirt.sh
    user_env/source/globus-user-env.csh
Delete:
    core/setup/globus-sh-tools.sh.in
    core/source/globus_config.h.in

+++ common/source/library/globus_error.c
@@ -237,7 +237,7 @@ globus_error_construct_base (globus_module_descriptor_t * source_module,
  **********************************************************************/
 
 static globus_object_cache_t s_result_to_object_mapper;
-static unsigned long         s_next_available_result_count;
+static uint32_t              s_next_available_result_count;
 static globus_mutex_t        s_result_to_object_mutex;
 
 static int  s_error_cache_initialized = 0;
